requirements.txt
Django==4.2.10
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
psycopg2-binary==2.9.10
python-decouple==3.8
celery==5.5.3
redis==6.4.0
gunicorn==21.2.0
whitenoise==6.6.0
firebase-admin==7.1.0
Pillow==10.4.0
requests==2.31.0
django-tenants==3.5.0 